+++ src/tests/plane.cpp
@@ -2,7 +2,7 @@
 #include "main.h"
 
 void BoxPermutationsRounded(struct notcurses* nc, struct ncplane* n, unsigned edges) {
-  int dimx, dimy;
+  unsigned dimx, dimy;
   ncplane_dim_yx(n, &dimy, &dimx);
   REQUIRE(2 < dimy);
   REQUIRE(47 < dimx);
@@ -38,9 +38,9 @@ TEST_CASE("Plane") {
 
   // Dimensions of the standard plane ought be the same as those of the context
   SUBCASE("StdPlaneDimensions") {
-    int cols, rows;
+    unsigned cols, rows;
     notcurses_term_dim_yx(nc_, &rows, &cols);
-    int ncols, nrows;
+    unsigned ncols, nrows;
     ncplane_dim_yx(n_, &nrows, &ncols);
     CHECK(rows == nrows);
     CHECK(cols == ncols);
@@ -48,7 +48,7 @@ TEST_CASE("Plane") {
 
   // Verify that we can move to all four coordinates of the standard plane
   SUBCASE("MoveStdPlaneDimensions") {
-    int cols, rows;
+    unsigned cols, rows;
     notcurses_term_dim_yx(nc_, &rows, &cols);
     CHECK(0 == ncplane_cursor_move_yx(n_, 0, 0));
     int x, y;
@@ -71,7 +71,7 @@ TEST_CASE("Plane") {
 
   // Verify that we can move to all four coordinates of the standard plane
   SUBCASE("MoveBeyondPlaneFails") {
-    int cols, rows;
+    unsigned cols, rows;
     notcurses_term_dim_yx(nc_, &rows, &cols);
     CHECK(0 > ncplane_cursor_move_yx(n_, -2, 0));
     CHECK(0 > ncplane_cursor_move_yx(n_, -2, -2));
@@ -198,13 +198,13 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("HorizontalLines") {
-    int x, y;
+    unsigned x, y;
     ncplane_dim_yx(n_, &y, &x);
     REQUIRE(0 < y);
     REQUIRE(0 < x);
     nccell c{};
     nccell_load(n_, &c, "-");
-    for(int yidx = 0 ; yidx < y ; ++yidx){
+    for(unsigned yidx = 0 ; yidx < y ; ++yidx){
       CHECK(0 == ncplane_cursor_move_yx(n_, yidx, 1));
       CHECK(x - 2 == ncplane_hline(n_, &c, x - 2));
       int posx, posy;
@@ -217,13 +217,13 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("VerticalLines") {
-    int x, y;
+    unsigned x, y;
     ncplane_dim_yx(n_, &y, &x);
     REQUIRE(0 < y);
     REQUIRE(0 < x);
     nccell c{};
     nccell_load(n_, &c, "|");
-    for(int xidx = 1 ; xidx < x - 1 ; ++xidx){
+    for(unsigned xidx = 1 ; xidx < x - 1 ; ++xidx){
       CHECK(0 == ncplane_cursor_move_yx(n_, 1, xidx));
       CHECK(y - 2 == ncplane_vline(n_, &c, y - 2));
       int posx, posy;
@@ -237,7 +237,7 @@ TEST_CASE("Plane") {
 
   // reject attempts to draw boxes beyond the boundaries of the ncplane
   SUBCASE("BadlyPlacedBoxen") {
-    int x, y;
+    unsigned x, y;
     ncplane_dim_yx(n_, &y, &x);
     REQUIRE(2 < y);
     REQUIRE(2 < x);
@@ -274,7 +274,7 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("BoxPermutationsDouble") {
-    int dimx, dimy;
+    unsigned dimx, dimy;
     ncplane_dim_yx(n_, &dimy, &dimx);
     REQUIRE(2 < dimx);
     REQUIRE(47 < dimx);
@@ -289,7 +289,7 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("PerimeterRoundedBox") {
-    int x, y;
+    unsigned x, y;
     ncplane_dim_yx(n_, &y, &x);
     REQUIRE(2 < y);
     REQUIRE(2 < x);
@@ -299,7 +299,7 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("PerimeterRoundedBoxSized") {
-    int x, y;
+    unsigned x, y;
     ncplane_dim_yx(n_, &y, &x);
     REQUIRE(2 < y);
     REQUIRE(2 < x);
@@ -309,7 +309,7 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("PerimeterDoubleBox") {
-    int x, y;
+    unsigned x, y;
     ncplane_dim_yx(n_, &y, &x);
     REQUIRE(2 < y);
     REQUIRE(2 < x);
@@ -319,7 +319,7 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("PerimeterDoubleBoxSized") {
-    int x, y;
+    unsigned x, y;
     ncplane_dim_yx(n_, &y, &x);
     REQUIRE(2 < y);
     REQUIRE(2 < x);
@@ -401,7 +401,7 @@ TEST_CASE("Plane") {
   // for the standard plane, and that we can change it.
   SUBCASE("UserPtr") {
     CHECK(nullptr == ncplane_userptr(n_));
-    int x, y;
+    unsigned x, y;
     void* sentinel = &x;
     notcurses_term_dim_yx(nc_, &y, &x);
     struct ncplane_options nopts = {
@@ -427,7 +427,7 @@ TEST_CASE("Plane") {
   // create a new plane, the same size as the terminal, and verify that it
   // occupies the same dimensions as the standard plane.
   SUBCASE("NewPlaneSameSize") {
-    int x, y;
+    unsigned x, y;
     notcurses_term_dim_yx(nc_, &y, &x);
     struct ncplane_options nopts = {
       .y = 0,
@@ -439,11 +439,11 @@ TEST_CASE("Plane") {
     };
     struct ncplane* ncp = ncplane_create(n_, &nopts);
     REQUIRE(ncp);
-    int px, py;
+    unsigned px, py;
     ncplane_dim_yx(ncp, &py, &px);
     CHECK(y == py);
     CHECK(x == px);
-    int sx, sy;
+    unsigned sx, sy;
     ncplane_dim_yx(n_, &sy, &sx);
     CHECK(sy == py);
     CHECK(sx == px);
@@ -453,7 +453,7 @@ TEST_CASE("Plane") {
   // create a new plane, the same size as the terminal, and verify that it
   // occupies the same dimensions as the standard plane, but on another pile.
   SUBCASE("NewPileSameSize") {
-    int x, y;
+    unsigned x, y;
     notcurses_term_dim_yx(nc_, &y, &x);
     struct ncplane_options nopts = {
       .y = 0,
@@ -465,11 +465,11 @@ TEST_CASE("Plane") {
     };
     struct ncplane* ncp = ncpile_create(nc_, &nopts);
     REQUIRE(ncp);
-    int px, py;
+    unsigned px, py;
     ncplane_dim_yx(ncp, &py, &px);
     CHECK(y == py);
     CHECK(x == px);
-    int sx, sy;
+    unsigned sx, sy;
     ncplane_dim_yx(n_, &sy, &sx);
     CHECK(sy == py);
     CHECK(sx == px);
@@ -484,7 +484,7 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("ShrinkPlane") {
-    int maxx, maxy;
+    unsigned maxx, maxy;
     int x = 0, y = 0;
     notcurses_term_dim_yx(nc_, &maxy, &maxx);
     struct ncplane_options nopts = {
@@ -528,9 +528,9 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("GrowPlane") {
-    int maxx = 2, maxy = 2;
+    unsigned maxx = 2, maxy = 2;
     int x = 0, y = 0;
-    int dimy, dimx;
+    unsigned dimy, dimx;
     notcurses_term_dim_yx(nc_, &dimy, &dimx);
     x = dimx / 2 - 1;
     y = dimy / 2 - 1;
@@ -552,7 +552,7 @@ TEST_CASE("Plane") {
       REQUIRE(0 == ncplane_resize(newp, 1, 1, maxy - 3, maxx - 3, 1, 1, maxy, maxx));
       // FIXME check dims, pos
     }
-    while(y < dimy){
+    while(y < static_cast<int>(dimy)){
       ++maxy;
       if(y){
         ++y;
@@ -560,7 +560,7 @@ TEST_CASE("Plane") {
       REQUIRE(0 == ncplane_resize(newp, 1, 0, maxy - 2, maxx, 1, 0, maxy, maxx));
       // FIXME check dims, pos
     }
-    while(x < dimx){
+    while(x < static_cast<int>(dimx)){
       ++maxx;
       if(x){
         ++x;
@@ -588,7 +588,7 @@ TEST_CASE("Plane") {
     CHECK(0 == testcell.stylemask);
     CHECK(0 == testcell.channels);
     nccell_release(n_, &testcell);
-    int dimy, dimx;
+    unsigned dimy, dimx;
     ncplane_dim_yx(n_, &dimy, &dimx);
     REQUIRE(0 == ncplane_cursor_move_yx(n_, 1, dimx - strlen(STR2)));
     REQUIRE(0 < ncplane_putstr(n_, STR2));
@@ -626,7 +626,7 @@ TEST_CASE("Plane") {
     CHECK(0 == testcell.stylemask);
     CHECK(0 == testcell.channels);
     nccell_release(n_, &testcell);
-    int dimy, dimx;
+    unsigned dimy, dimx;
     ncplane_dim_yx(n_, &dimy, &dimx);
     REQUIRE(0 == ncplane_cursor_move_yx(n_, 1, dimx - mbstowcs(nullptr, STR2, 0)));
     REQUIRE(0 < ncplane_putstr(n_, STR2));
@@ -684,7 +684,7 @@ TEST_CASE("Plane") {
 
   SUBCASE("BoxGradients") {
     const auto sidesz = 5;
-    int dimx, dimy;
+    unsigned dimx, dimy;
     ncplane_dim_yx(n_, &dimy, &dimx);
     REQUIRE(20 < dimy);
     REQUIRE(40 < dimx);
@@ -722,7 +722,7 @@ TEST_CASE("Plane") {
 
   SUBCASE("BoxSideColors") {
     const auto sidesz = 5;
-    int dimx, dimy;
+    unsigned dimx, dimy;
     ncplane_dim_yx(n_, &dimy, &dimx);
     REQUIRE(20 < dimy);
     REQUIRE(40 < dimx);
@@ -778,14 +778,14 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("NewPlaneOnRight") {
-    int ncols, nrows;
+    unsigned ncols, nrows;
     ncplane_dim_yx(n_, &nrows, &ncols);
     nccell ul{}, ll{}, lr{}, ur{}, hl{}, vl{};
     int y, x;
     ncplane_yx(n_, &y, &x);
     struct ncplane_options nopts = {
       .y = y,
-      .x = ncols - 3,
+      .x = static_cast<int>(ncols) - 3,
       .rows = 2,
       .cols = 2,
       .userptr = nullptr, .name = nullptr, .resizecb = nullptr, .flags = 0,
@@ -801,7 +801,7 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("MoveToLowerRight") {
-    int ncols, nrows;
+    unsigned ncols, nrows;
     ncplane_dim_yx(n_, &nrows, &ncols);
     nccell ul{}, ll{}, lr{}, ur{}, hl{}, vl{};
     int y, x;
@@ -859,7 +859,7 @@ TEST_CASE("Plane") {
   }
 
   SUBCASE("MouseEvent") {
-    int dimy, dimx;
+    unsigned dimy, dimx;
     notcurses_stddim_yx(nc_, &dimy, &dimx);
     struct ncplane_options nopts = {
       .y = 1,
@@ -1087,14 +1087,14 @@ TEST_CASE("Plane") {
     int y, x;
     CHECK(1 == ncplane_set_base(n, " ", 0, channels));
     CHECK(0 == notcurses_render(nc_));
-    for(int i = 0 ; i < ncplane_dim_y(n) ; ++i){
+    for(unsigned i = 0 ; i < ncplane_dim_y(n) ; ++i){
       ncplane_cursor_yx(n, &y, &x);
       CHECK(i == y);
       CHECK(0 == x);
       CHECK(0 < ncplane_putstr(n, "here's a line\n"));
       CHECK(0 == notcurses_render(nc_));
     }
-    for(int i = 0 ; i < ncplane_dim_y(n) ; ++i){
+    for(unsigned i = 0 ; i < ncplane_dim_y(n) ; ++i){
       ncplane_cursor_yx(n, &y, &x);
       CHECK(ncplane_dim_y(n) - 1 == y);
       CHECK(0 == x);
